+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to Stability Matrix will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html).
 
+## v2.11.1
+### Added
+- Added Rename option back to the Checkpoints page
+### Changed
+- Unobserved Task Exceptions across the app will now show a toast notification to aid in debugging
+### Fixed
+- Fixed [#689](https://github.com/LykosAI/StabilityMatrix/issues/689) - New ComfyUI installs encountering launch error due to torch 2.0.0 update, added pinned `numpy==1.26.4` to install and update.
+- Fixed Inference image mask editor's 'Load Mask' not able to load image files
+- Fixed Fooocus ControlNet default config shared folder mode not taking effect
+
 ## v2.11.0
 ### Added
 #### Packages